Define Webhook based transport
This transport proposal aims to optimize Server to Server communication (hosted clients), especially around long-running sessions and operations.

To make MCP work for long running tasks like deep analysis, this proposal defines a transport that leverages http webhooks for bidirectionality.

So whilst the SSE transport enables remote servers, a webhook based approach enables remoteness without the need of long-lived open socket connections.

As an extension of this transport, we could introduce an (optional) SSE endpoint for message streaming for reduced latency. However, it is critical to not bind the lifetime of the session to the request lifetime.