Allow `Implementation`s to optionally expose a `websiteUrl` and `iconUrl`
This proposed update to the specification allows MCP
`Implementation`s to optionally declare a `websiteUrl` and
`iconUrl`.

Most MCP clients will render custom UI showing when an MCP
server is used - for example, when one of the server's tools
is invoked.

Existing metadata allows clients to show a human-readable name
for the tool (`ToolAnnotations.title`) and a name for the server
`Implementation.name`).

These changes will allow clients to render richer UI, showing
a logo for the server (e.g. the GitHub logo for the
[GitHub MCP server][1]) and linking to the server's website
for more information.

[1]: https://github.com/github/github-mcp-server