Add `Resource.uriScope` to enable direct reads
When reading a binary resource via `resources/read`, the resource
content must be encoded into Base64 and wrapped in JSON on the server,
and then must be decoded on the client.  Base64 increases payload size
by 33%, so this process can add significant overhead for large
resources.

This commit adds an optional `uriScope` property to `Resource`, which
indicates when the resource content may be read directly via its URI,
thereby avoiding the overhead.  If the value of `uriScope` is
`"external"`, clients may attempt to read via the URI.  If the value of
`uriScope` is `"internal"`, clients must verify that the server is local
before attempting to read via the URI.  If these conditions are not met
or if the read fails, clients should fall back to `resources/read`.

It's also worth mentioning that if a client SDK's `resources/read` API
method accepts a resource object (as returned by `resources/list`), then
the API method can transparently take advantage of this optimization.