Change link underline color
Mintlify uses the theme's "light" color as the underline color for links
in both light mode and dark mode.  Prior to this commit, our "light"
color was set to `#FAFAFA`, which is indistinguishable from the
background color in light mode, `#FFFFFF`.  Thus, in light mode, links
were indistinguishable from bolded text.

To address the issue, this commit changes our "light" color to
`#607F9F`.  The color was chosen by desaturating the default "light"
color of the Willow theme, `#007FFF`:

    #007FFF == hsl(210, 100%, 50%) => hsl(210, 25%, 50%) == #607F9F