Clarify tool error categorization
Prior to this commit, the tools doc listed "Invalid arguments" as an
example of a protocol error.  This was likely meant to refer to the
arguments of the JSON-RPC request itself (i.e. `params`), rather than
arguments of the tool (i.e. `params.arguments`).

This commit clarifies the wording as "Invalid request params", adds
associated error codes (to match [`prompts.mdx`][]), and improves the
section's formatting.

[`prompts.mdx`]: https://github.com/modelcontextprotocol/modelcontextprotocol/blob/8fe803b9085ed94be7dad5e4379b5fe773c4922e/docs/specification/draft/server/prompts.mdx?plain=1#L257-L259

Closes #290.

Co-authored-by: Mandeep Cheema <mandeep@cheema.uk>